requirements.txt
amqp==5.1.1
asgiref==3.4.1
astroid==2.11.7
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dill==0.3.4
Django==3.2.15
django-cors-headers==3.10.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
phonenumbers==8.12.53
platformdirs==2.4.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pylint==2.13.9
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
